files/en-us/mozilla/qa/chrome_tests/raw.html

<h3 id="Introduction" name="Introduction">Introduction</h3>

<p>A chrome test is basically a <a href="/en-US/docs/Mochitest" title="en/Mochitest">Mochitest</a> running with chrome privileges.</p>

<p>The chrome test suite is an automated testing framework designed to allow testing of application chrome windows using JavaScript. It currently allows you to run JavaScript code in the same scope as the main Firefox browser window with chrome privileges and reports results using the same functions as the <a href="/en-US/docs/Mochitest" title="en/Mochitest">Mochitest test framework</a>. The chrome test suite depends on runtests.py from the Mochitest framework.</p>

<h3 id="Running_the_chrome_tests" name="Running_the_chrome_tests">Running the chrome tests</h3>

<p>To run Mochitest you need to <a href="/en-US/docs/Developer_Guide/Build_Instructions" title="en/Build_Documentation">build Mozilla</a> with your changes and run the command:</p>

<pre>./mach mochitest -f chrome
</pre>

<p>To run a single test, just pass the path to the test into mach:</p>

<pre>./mach mochitest -f chrome toolkit/content/tests/chrome/test_largemenu.xul</pre>

<p>You can also pass the path to a directory containing many tests. Run <strong>mach help mochitest</strong> for full documentation.</p>

<h3 id="Writing_chrome_tests" name="Writing_chrome_tests">Writing chrome tests</h3>

<p>A chrome tests is basically a <a href="/en-US/docs/Mochitest" title="en/Mochitest">Mochitest</a> running with chrome privileges, i.e. code and UI are referenced by <code>chrome://</code> URIs. A basic XUL test file could look like this:</p>

<pre class="eval">&lt;?xml version="1.0"?&gt;
&lt;?xml-stylesheet href="chrome://global/skin" type="text/css"?&gt;
&lt;?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?&gt;

&lt;window title="Demo Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="RunTest();"&gt;
  &lt;title&gt;Demo Test&lt;/title&gt;

  &lt;script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/&gt;

  &lt;script type="application/javascript"&gt;
  &lt;![CDATA[
    SimpleTest.waitForExplicitFinish();

    function RunTest()
    {
      ok  (true ==  1, "this passes");
      todo(true === 1, "this fails");
      SimpleTest.finish();
    }
  ]]&gt;
  &lt;/script&gt;

  &lt;body xmlns="http://www.w3.org/1999/xhtml"&gt;
    &lt;p id="display"&gt;&lt;/p&gt;
    &lt;div id="content" style="display: none"&gt;&lt;/div&gt;
    &lt;pre id="test"&gt;&lt;/pre&gt;
  &lt;/body&gt;
&lt;/window&gt;
</pre>

<p>The "RunTest" function is invoked by the test's <code>onload</code> handler, not by the test harness.</p>

<p>The comparison functions are identical to those supported by Mochitests, see <a href="/en-US/docs/Mochitest#How_do_the_comparison_functions_work.3F" title="en/Mochitest#How_do_the_comparison_functions_work.3F">how the comparison functions work</a> in the Mochitest documentation for more details. The <a class="external" href="http://lxr.mozilla.org/mozilla/source/testing/mochitest/tests/SimpleTest/EventUtils.js">EventUtils helper functions</a> are available on the "EventUtils" object defined in the global scope.</p>

<p>The test suite also supports asynchronous tests, using the same function names as Mochitest. Call SimpleTest.waitForExplicitFinish() if you want to delay reporting a result until after RunTest() has returned. Call SimpleTest.finish() once the test is complete. Be aware that the test harness will mark tests that take too long to complete as FAILED (the current timeout is 15 seconds).</p>

<p>Any exceptions thrown while running a test will be caught and reported in the test output as a failure. Exceptions thrown outside of the test's context (e.g. in a timeout, event handler, etc) will not be caught, but will result in a timed out test if they prevent finish() from being called.</p>

<p>The test file name must be prefixed with "test_", and must have a file extension of ".xul". Files that don't match this pattern will be ignored by the test harness, but you still can include them. For example, a XUL window file opened by your test_demo.xul via openDialog should be named window_demo.xul. Putting the bug number in the file name is recommended if your test verifies a bugfix, e.g. "test_bug123456.xul".</p>

<p>Helper files can be included, for example, from <code>https://example.com/chrome/dom/workers/test/serviceworkers/serviceworkermanager_iframe.html</code>.</p>

<h3 id="Adding_a_new_chrome_test_to_the_tree" name="Adding_a_new_chrome_test_to_the_tree">Adding a new chrome test to the tree</h3>

<p>To add a new chrome test to the tree, follow the <a href="/en-US/docs/Mochitest#Adding_new_Mochitest_tests_to_the_tree" title="en/Mochitest#Adding_new_Mochitest_tests_to_the_tree">Mochitest instructions</a>, keeping in mind that the chrome tests must be copied into _tests/testing/mochitest/chrome instead of _tests/testing/mochitest/mochitest and use a manifest named chrome.ini instead of mochitest.ini.</p>
